+++ table/persistent_cache_helper.cc
@@ -2,7 +2,9 @@
 //  This source code is licensed under the BSD-style license found in the
 //  LICENSE file in the root directory of this source tree. An additional grant
 //  of patent rights can be found in the PATENTS file in the same directory.
+
 #include "table/persistent_cache_helper.h"
+#include "table/block_based_table_reader.h"
 #include "table/format.h"
 
 namespace rocksdb {
