+++ table/persistent_cache_options.h
@@ -1,0 +1,34 @@
+//  Copyright (c) 2011-present, Facebook, Inc.  All rights reserved.
+//  This source code is licensed under the BSD-style license found in the
+//  LICENSE file in the root directory of this source tree. An additional grant
+//  of patent rights can be found in the PATENTS file in the same directory.
+#pragma once
+
+#include <string>
+
+#include "include/rocksdb/persistent_cache.h"
+#include "util/statistics.h"
+
+namespace rocksdb {
+
+// PersistentCacheOptions
+//
+// This describe the caching behavior for page cache
+// This is used to pass the context for caching and the cache handle
+struct PersistentCacheOptions {
+  PersistentCacheOptions() {}
+  explicit PersistentCacheOptions(
+      const std::shared_ptr<PersistentCache>& _persistent_cache,
+      const std::string _key_prefix, Statistics* const _statistics)
+      : persistent_cache(_persistent_cache),
+        key_prefix(_key_prefix),
+        statistics(_statistics) {}
+
+  virtual ~PersistentCacheOptions() {}
+
+  std::shared_ptr<PersistentCache> persistent_cache;
+  std::string key_prefix;
+  Statistics* statistics = nullptr;
+};
+
+}  // namespace rocksdb
